+++ source/fonctions/rk4.h
@@ -6,37 +6,37 @@ void rk4(vec2 M[RES][RES])
     {
         for (size_t j = 0; j < RES; j++)
         {
-            float px = M[i][j].x;
-            float py = M[i][j].y;
-            float vx = v_0_x;
-            float vy = v_0_y;
+            double px = M[i][j].x;
+            double py = M[i][j].y;
+            double vx = v_0_x;
+            double vy = v_0_y;
 
-            for (float t = 0; t < t_f; t += dt)
+            for (double t = 0; t < t_f; t += dt)
             {
                 if (condition_capture(px, py))
                     break;
 
-                // k1 
+                // k1
                 vec2 a1 = accel(px, py);
-                float k1_x = vx, k1_y = vy;       // vitesse initiale
-                float k1_vx = a1.x, k1_vy = a1.y; // accel initiale
+                double k1_x = vx, k1_y = vy;       // vitesse initiale
+                double k1_vx = a1.x, k1_vy = a1.y; // accel initiale
 
                 // k2  en utilisant k1
                 vec2 a2 = accel(px + k1_x * dt / 2, py + k1_y * dt / 2);
-                float k2_x = vx + k1_vx * dt / 2, k2_y = vy + k1_vy * dt / 2;
-                float k2_vx = a2.x, k2_vy = a2.y;
+                double k2_x = vx + k1_vx * dt / 2, k2_y = vy + k1_vy * dt / 2;
+                double k2_vx = a2.x, k2_vy = a2.y;
 
                 // k3  en utilisant k2
                 vec2 a3 = accel(px + k2_x * dt / 2, py + k2_y * dt / 2);
-                float k3_x = vx + k2_vx * dt / 2, k3_y = vy + k2_vy * dt / 2;
-                float k3_vx = a3.x, k3_vy = a3.y;
+                double k3_x = vx + k2_vx * dt / 2, k3_y = vy + k2_vy * dt / 2;
+                double k3_vx = a3.x, k3_vy = a3.y;
 
                 // k4 en utilisant k3
                 vec2 a4 = accel(px + k3_x * dt, py + k3_y * dt);
-                float k4_x = vx + k3_vx * dt, k4_y = vy + k3_vy * dt;
-                float k4_vx = a4.x, k4_vy = a4.y;
+                double k4_x = vx + k3_vx * dt, k4_y = vy + k3_vy * dt;
+                double k4_vx = a4.x, k4_vy = a4.y;
 
-                //m.a.j des positions et vitesses formule : (1/6 * k1) + (2/6 * k2) + (2/6 * k3) + (1/6 * k4)
+                // m.a.j des positions et vitesses formule : (1/6 * k1) + (2/6 * k2) + (2/6 * k3) + (1/6 * k4)
                 px += dt / 6.0f * (k1_x + 2 * k2_x + 2 * k3_x + k4_x);
                 py += dt / 6.0f * (k1_y + 2 * k2_y + 2 * k3_y + k4_y);
                 vx += dt / 6.0f * (k1_vx + 2 * k2_vx + 2 * k3_vx + k4_vx);
